[62-JavaCrypto-AES] 선택가능하도록 변경 - 프로덕션 코드
