+++ member-context/src/main/java/com/membercontext/memberAPI/infrastructure/encryption/aes256/exception/JavaCryptoException.java
@@ -1,0 +1,17 @@
+package com.membercontext.memberAPI.infrastructure.encryption.aes256.exception;
+
+
+import com.membercontext.memberAPI.infrastructure.encryption.aes256.exception.ErrorCode.JavaCryptoErrorCode;
+import lombok.Getter;
+
+@Getter
+
+public class JavaCryptoException extends RuntimeException {
+    private final JavaCryptoErrorCode javaCryptoCode;
+
+    public JavaCryptoException(JavaCryptoErrorCode javaCryptoCode) {
+        super(javaCryptoCode.getDeatil());
+        this.javaCryptoCode = javaCryptoCode;
+    }
+}
+
