--- member-context/src/main/java/com/membercontext/memberAPI/infrastructure/encryption/db/JavaCryptoIVRepository.java
@@ -1,14 +1,0 @@
-package com.membercontext.memberAPI.infrastructure.encryption.db;
-
-
-
-
-import com.membercontext.memberAPI.infrastructure.encryption.db.vo.JavaCryptoIV;
-
-import java.util.Optional;
-
-public interface JavaCryptoIVRepository {
-
-    public void save(JavaCryptoIV javaCryptoIV);
-    public Optional<JavaCryptoIV> get(Long id);
-}
